Move data-utils and clean up syntax a bit
